requirements.txt
fastmcp
numpy
pyaudio
requests
python-dotenv
google-genai
reachy-mini[gstreamer]
pyttsx3
opencv-contrib-python